requirements.txt
numpy==1.20.2
nltk==3.6.2
ibm_watson==5.2.2
ibm_cloud_sdk_core==3.10.1
wordcloud==1.8.1
psaw==0.1.0
pandas==1.2.4
tweepy==3.10.0
tqdm==4.60.0
Werkzeug==0.16.0
Pillow==8.3.1
scikit_learn==0.24.2
